+++ security-service/src/main/java/com/gitlab/util/JwtUtil.java
@@ -1,0 +1,64 @@
+package com.gitlab.util;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import java.security.Key;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Function;
+
+@Component
+public class JwtUtil {
+
+    @Value("${jwt_secret}")
+    private String secret;
+
+    public void validateToken(final String token) {
+        Jwts.parserBuilder().setSigningKey(getSigningKey()).build().parseClaimsJws(token.substring(7));
+    }
+
+    public String generateToken(String userEmail) {
+        Map<String, Object> claims = new HashMap<>();
+        return createToken(claims, userEmail);
+    }
+
+    private String createToken(Map<String, Object> claims, String userEmail) {
+        return Jwts.builder()
+                .setClaims(claims)
+                .setSubject(userEmail)
+                .setIssuedAt(new Date(System.currentTimeMillis()))
+                .setExpiration(new Date(System.currentTimeMillis() + 1000 * 60 * 30))
+                .signWith(getSigningKey(), SignatureAlgorithm.HS256).compact();
+    }
+
+    private Key getSigningKey() {
+        var keyBytes = Decoders.BASE64.decode(secret);
+        return Keys.hmacShaKeyFor(keyBytes);
+    }
+
+    public String extractUsername(String token) {
+        return extractClaims(token, Claims::getSubject);
+    }
+
+    public <T> T extractClaims(String token, Function<Claims, T> claimsResolver) {
+        final var claims = extractAllClaims(token);
+        return claimsResolver.apply(claims);
+    }
+
+    private Claims extractAllClaims(String token) {
+        return Jwts
+                .parserBuilder()
+                .setSigningKey(getSigningKey())
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+}