+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 - #83 - Generate test data for PostomatRestControllerIT
 - #12 - Order entity
 - #61 - Front CRUD for Passport
-- #101 - added fuzzy search for product
+- #92 - Security service
+- #101 - Add fuzzy search for product
 
 ## [1.1.0] - 8.09.2023
 - #5 - added user entity